+++ sdk/src/main/java/com/walmartlabs/concord/sdk/SecretStoreService.java
@@ -22,7 +22,7 @@ package com.walmartlabs.concord.sdk;
 
 public interface SecretStoreService {
 
-    Secret fetch(String instanceId, String secretName, String password) throws ClientException;
+    Secret fetch(String instanceId, String orgName, String secretName, String password) throws ClientException;
 
     String decryptString(String instanceId, String s) throws ClientException;
 }